Other(s)

+++ lot/extend/shield/engine/plug/shield.php
@@ -1,14 +1,16 @@
 <?php
 
-Shield::_('message', function(string $kin = "") {
-    if ($message = Message::get($kin, false)) {
-        $message = str_replace([
-            '<message type="',
-            '</message>'
-        ], [
-            '<p class="message message-',
-            '</p>'
-        ], $message);
-        echo '<div class="messages p">' . $message . '</div>';
-    }
-});
+if (Extend::exist('message')) {
+    Shield::_('message', function(string $kin = "") {
+        if ($message = Message::get($kin, false)) {
+            $message = str_replace([
+                '<message type="',
+                '</message>'
+            ], [
+                '<p class="message message-',
+                '</p>'
+            ], $message);
+            echo '<div class="messages p">' . $message . '</div>';
+        }
+    });
+}
